convert TableTest addMethods() calls
