+++ tests/TestCase/Mailer/MailerDefaultProfileRestorationTest.php
@@ -37,9 +37,11 @@ class MailerDefaultProfileRestorationTest extends TestCase
     }
 }
 
+// phpcs:disable
 class CustomMailer extends Mailer
 {
     public function test($to, $subject)
     {
     }
 }
+// phpcs:enable
