+++ tests/TestCase/Mailer/MailerSendWithUnsetTemplateDefaultsToActionNameTest.php
@@ -36,9 +36,11 @@ class MailerSendWithUnsetTemplateDefaultsToActionNameTest extends TestCase
     }
 }
 
+// phpcs:disable
 class CustomMailer extends Mailer
 {
     public function test($to, $subject)
     {
     }
 }
+// phpcs:enable
